--- apps/bot/jest.config.cjs
@@ -1,9 +1,0 @@
-// @ts-check
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
-	preset: "ts-jest",
-	testEnvironment: "node",
-	clearMocks: true,
-	resetMocks: true,
-	testPathIgnorePatterns: ["/node_modules/", "/dist/"],
-};
