+++ apps/bot/src/commands/log.ts
@@ -1,0 +1,1 @@
+// TODO: Choose channel where events post to
