+++ apps/bot/src/lib/parser/issue-alert.ts
@@ -25,7 +25,3 @@ export async function getIssueAlertEmbed(json: Record<string, any>) {
 
 	return embed;
 }
-
-export function getIssueAlertUrl(json: Record<string, any>) {
-	return json?.data?.event?.url as string | undefined;
-}
