+++ apps/bot/src/listeners/unfurler.ts
@@ -2,13 +2,16 @@ import { CommandArgs } from "../typedefs";
 
 // TODO: listen for sentry links and post basic information
 export default async ({ bot }: CommandArgs) => {
+	bot.on("message", (message) => {
+		const matches = message.cleanContent.matchAll(
+			/sentry\.io\/organizations\/(?<organization>\w+)\/issues\/(?<issue>\d+)/g
+		);
 
-	// TODO: impl
-	bot.on("interaction", () => {
-		// check to see if message contains link
-		// check to see if preview is on for guild/project
-		// get information about issue
-		// post embed as reply
-		console.log("interaction");
+		if (matches != null) {
+			console.log("Unfurl");
+			// TODO: check to see if preview is on for guild/project
+			// TODO: get information about issue
+			// TODO: post embed as reply
+		}
 	});
 };
