+++ package.json
@@ -28,7 +28,7 @@
 	"devDependencies": {
 		"@remix-run/dev": "^1.1.3",
 		"@types/debug": "^4.1.7",
-		"@types/jest": "^27.4.0",
+		"@types/ioredis": "^4.28.10",
 		"@types/node": "^17.0.13",
 		"@types/react": "^17.0.24",
 		"@types/react-dom": "^17.0.9",
@@ -41,16 +41,15 @@
 		"eslint-config-xo": "^0.39.0",
 		"eslint-config-xo-typescript": "^0.49.0",
 		"eslint-import-resolver-typescript": "^2.5.0",
-		"jest": "^27.5.1",
 		"postcss": "^8.4.6",
 		"prettier": "^2.5.1",
 		"prisma": "^3.9.2",
 		"rimraf": "^3.0.2",
 		"tailwindcss": "^3.0.22",
-		"ts-jest": "^27.1.3",
 		"ts-node": "^10.4.0",
 		"turbo": "^1.1.2",
-		"typescript": "^4.5.5"
+		"typescript": "^4.5.5",
+		"vitest": "^0.12.6"
 	},
 	"dependencies": {
 		"@prisma/client": "^3.9.2"
