+++ packages/sentry-parser/README.md
@@ -1,0 +1,3 @@
+# sentry-parser
+
+WIP library to parse Sentry webhooks
