+++ packages/sentry-parser/src/installation.ts
