+++ packages/sentry-parser/src/issue-alert.ts
@@ -1,0 +1,5 @@
+import { MessageEmbed } from "discord.js";
+
+export function getIssueAlertUrl(json: Record<string, any>) {
+	return json?.data?.event?.url as string | undefined;
+}
