+++ packages/sentry-parser/test/example.test.ts
@@ -1,3 +1,5 @@
+import { test, expect } from "vitest";
+
 test("succeeds", () => {
 	expect(true).toBe(true);
 });
