+++ packages/sentry-parser/test/mocks/installations/example.json
