+++ packages/sentry-parser/test/mocks/issue-alert/example.json
