+++ packages/sentry-parser/test/mocks/issues/example.json
