+++ packages/sentry-parser/test/mocks/metric-alert/example.json
