+++ packages/sentry-parser/test/mocks/uninstallations/example.json
